feat: add app observers
