+++ lib/app/logger.dart
@@ -1,0 +1,3 @@
+import 'package:logger/logger.dart';
+
+final logger = Logger(printer: PrettyPrinter());
